Fixed rts usage on tests

+++ tests/test_common.py
@@ -172,7 +172,15 @@ def test_common_exporter_logic():
     # negative test
     common_exporter_logic(None, None, None, None, None, None, None, None, None, None)
     try:
-        rts = redis.Redis(port=16379)
+        import os
+        # Ensure we have the test DB to store results
+        assert "RTS_PORT" in os.environ
+        rts_port = os.environ.get("RTS_PORT",None)
+        rts_host = os.getenv("RTS_DATASINK_HOST", None)
+        rts_pass = ""
+        if rts_host is None:
+            return
+        rts = redis.Redis(port=rts_port, host=rts_host)
         rts.ping()
         with open(
             "./tests/test_data/redis-benchmark-full-suite-1Mkeys-100B.yml", "r"
@@ -508,9 +516,16 @@ def test_dbconfig_keyspacelen_check():
     from redis import StrictRedis
     from redis.exceptions import ConnectionError
 
-    redis_port = 16379
+    import os
+    # Ensure we have the test DB to store results
+    assert "RTS_PORT" in os.environ
+    rts_port = os.environ.get("RTS_PORT",None)
+    rts_host = os.getenv("RTS_DATASINK_HOST", None)
+    rts_pass = ""
+    if rts_host is None:
+        return
     try:
-        redis = StrictRedis(port=redis_port)
+        redis = redis.Redis(port=rts_port, host=rts_host)
         redis.ping()
         redis.flushall()
         redis_conns = [redis]
@@ -582,9 +597,13 @@ def test_execute_init_commands():
     from redis import StrictRedis
     from redis.exceptions import ConnectionError
 
-    redis_port = 16379
+    import os
+    # Ensure we have the test DB to store results
+    assert "RTS_PORT" in os.environ
+    rts_port = os.environ.get("RTS_PORT",None)
+    
     try:
-        redis = StrictRedis(port=redis_port)
+        redis = StrictRedis(port=rts_port)
         redis.ping()
         redis.flushall()
         redis.config_resetstat()
